keep registry data all in one place
