+++ src/openfecli/commands/test.py
@@ -1,5 +1,4 @@
 import os
-import pathlib
 import sys
 
 import click
@@ -7,61 +6,8 @@ import pooch
 import pytest
 
 from openfecli import OFECommandPlugin
+from openfecli.data import _downloader
 from openfecli.utils import POOCH_CACHE, write
-
-
-def retrieve_all_test_data(path):
-    downloader = pooch.DOIDownloader(progressbar=True)
-
-    zenodo_cmet_data = dict(
-        base_url="doi:10.5281/zenodo.15200083/",
-        fname="cmet_results.tar.gz",
-        known_hash="md5:a4ca67a907f744c696b09660dc1eb8ec",
-    )
-    zenodo_rbfe_serial_data = dict(
-        base_url="doi:10.5281/zenodo.15042470/",
-        fname="rbfe_results_serial_repeats.tar.gz",
-        known_hash="md5:2355ecc80e03242a4c7fcbf20cb45487",
-    )
-    zenodo_rbfe_parallel_data = dict(
-        base_url="doi:10.5281/zenodo.15042470/",
-        fname="rbfe_results_parallel_repeats.tar.gz",
-        known_hash="md5:ff7313e14eb6f2940c6ffd50f2192181",
-    )
-    zenodo_abfe_data = dict(
-        base_url="doi:10.5281/zenodo.17348229/",
-        fname="abfe_results.zip",
-        known_hash="md5:547f896e867cce61979d75b7e082f6ba",
-    )
-    zenodo_septop_data = dict(
-        base_url="doi:10.5281/zenodo.17435569/",
-        fname="septop_results.zip",
-        known_hash="md5:2cfa18da59a20228f5c75a1de6ec879e",
-    )
-
-    def _infer_processor(fname: str):
-        if fname.endswith("tar.gz"):
-            return pooch.Untar()
-        elif fname.endswith("zip"):
-            return pooch.Unzip()
-        else:
-            return None
-
-    for d in [
-        zenodo_cmet_data,
-        zenodo_rbfe_serial_data,
-        zenodo_rbfe_parallel_data,
-        zenodo_abfe_data,
-        zenodo_septop_data,
-    ]:
-        pooch.retrieve(
-            url=d["base_url"] + d["fname"],
-            known_hash=d["known_hash"],
-            fname=d["fname"],
-            processor=_infer_processor(d["fname"]),
-            downloader=downloader,
-            path=path,
-        )
 
 
 @click.command("test", short_help="Run the OpenFE test suite")
@@ -86,7 +32,7 @@ def test(long, download_only):
     """
 
     if download_only:
-        retrieve_all_test_data(POOCH_CACHE)
+        _downloader.retrieve_all_test_data(POOCH_CACHE)
         sys.exit(0)
 
     try:
